book-component-added
