+++ backend/package.json
@@ -8,6 +8,7 @@
     "mongoose": "^6.6.0",
     "mongoose-increment": "^2.0.0",
     "morgan": "^1.10.0",
-    "multer": "^1.4.5-lts.1"
+    "multer": "^1.4.5-lts.1",
+    "nodemon": "^2.0.20"
   }
 }
