+++ backend/routes/books.js
@@ -8,6 +8,37 @@ const path = require("path");
 // route.post("/:id", (req, res) => {
 //   const id = req.params.id;
 // });
+function escapeRegex(text) {
+  return text.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&");
+}
+
+route.post("/get_books_by_name", (req, res) => {
+  if (req.body.name) {
+    const regex = new RegExp(escapeRegex(req.body.name), "gi");
+    console.log("reg:" + regex);
+    Book.find({ name: regex }, (err, books) => {
+      if (err) {
+        console.log(err);
+      } else {
+        return res.send(books);
+      }
+    });
+  }
+});
+
+route.post("/get_books_by_author", async (req, res) => {
+  if (req.body.author) {
+    const regex = new RegExp(escapeRegex(req.body.author), "gi");
+    console.log("reg:" + regex);
+    Book.find({ author: regex }, (err, books) => {
+      if (err) {
+        console.log(err);
+      } else {
+        return res.send(books);
+      }
+    });
+  }
+});
 
 route.post("/", async (req, res) => {
   // let book = Book.find({});
@@ -27,6 +58,7 @@ route.post("/create", upload.single("photo"), async (req, res) => {
     rating: req.body.rating,
     InStock: req.body.InStock,
     tags: req.body.tags,
+    price: req.body.price,
   });
   if (req.file) {
     newone.photo = req.file.path;
@@ -50,22 +82,13 @@ route.get("/all", async (req, res) => {
   return res.send(allbooks);
 });
 
-route.post("/get_books_by_name", async (req, res) => {
-  let books_by_name = await Book.find({
-    // name: { $regex: req.body.name, $opinions: "i" },
-    name: req.body.name,
-    // author: { $regex: req.body.author, $opinions: "i" },
-  });
-  return res.send(books_by_name);
-});
-
-route.post("/get_books_by_author", async (req, res) => {
-  let books_by_name = await Book.find({
-    // name: { $regex: req.body.name, $opinions: "i" },
-    author: req.body.author,
-    // author: { $regex: req.body.author, $opinions: "i" },
-  });
-  return res.send(books_by_name);
-});
+// route.post("/get_books_by_name", async (req, res) => {
+//   let books_by_name = await Book.find({
+//     // name: { $regex: req.body.name, $opinions: "i" },
+//     name: { $regex: req.body.name },
+//     // author: { $regex: req.body.author, $opinions: "i" },
+//   });
+//   return res.send(books_by_name);
+// });
 
 module.exports = route;
