+++ frontend/src/components/BookPopup.jsx
